Moved litter input chemistry calculation to inputs submodule to avoid circular import
