+++ tests/models/litter/test_chemistry.py
@@ -30,6 +30,7 @@ def test_calculate_new_pool_chemistries(
     litter_chemistry,
     litter_losses,
     post_consumption_pools,
+    input_chemistries,
 ):
     """Test that function to calculate updated pool chemistries works correctly."""
 
@@ -52,6 +53,7 @@ def test_calculate_new_pool_chemistries(
     actual_chemistries = litter_chemistry.calculate_new_pool_chemistries(
         litter_inputs=litter_inputs,
         litter_losses=litter_losses,
+        input_chemistries=input_chemistries,
         original_pools=post_consumption_pools,
         update_interval=2.0,
     )
@@ -207,201 +209,3 @@ def test_calculate_new_c_p_ratios(
 
     for key in actual_change.keys():
         assert np.allclose(actual_change[key], expected_change[key])
-
-
-def test_calculate_input_chemistries(litter_inputs):
-    """Check that calculation of input chemistries is correct."""
-    from dataclasses import asdict
-
-    from virtual_ecosystem.models.litter.chemistry import calculate_input_chemistries
-
-    expected_chemistries = {
-        "woody_lignin": [0.233, 0.545, 0.612, 0.378],
-        "above_structural_lignin": [0.25011178, 0.25345463, 0.54339369, 0.61992378],
-        "below_structural_lignin": [0.48590258, 0.56412613, 0.54265483, 0.67810978],
-        "woody_nitrogen": [60.7, 57.9, 73.1, 55.1],
-        "below_metabolic_nitrogen": [
-            20.32269136,
-            22.96676383,
-            26.06473456,
-            19.59251036,
-        ],
-        "below_structural_nitrogen": [
-            101.61345679,
-            114.83381916,
-            130.32367278,
-            97.96255178,
-        ],
-        "above_metabolic_nitrogen": [12.540983, 21.600478, 20.237902, 15.403147],
-        "above_structural_nitrogen": [62.91002, 110.3194, 109.3635, 75.59183],
-        "woody_phosphorus": [856.5, 675.4, 933.2, 888.8],
-        "below_metabolic_phosphorus": [
-            440.4591226,
-            226.94788998,
-            263.23576031,
-            196.40039357,
-        ],
-        "below_structural_phosphorus": [
-            2202.29561299,
-            1134.7394499,
-            1316.17880156,
-            982.00196785,
-        ],
-        "above_metabolic_phosphorus": [286.886303, 107.015923, 241.802298, 136.049497],
-        "above_structural_phosphorus": [
-            1488.595406,
-            580.6433876,
-            1408.378272,
-            610.0666667,
-        ],
-    }
-
-    actual_chemistries = calculate_input_chemistries(
-        litter_inputs=litter_inputs,
-        struct_to_meta_nitrogen_ratio=LitterConsts.structural_to_metabolic_n_ratio,
-        struct_to_meta_phosphorus_ratio=LitterConsts.structural_to_metabolic_p_ratio,
-    )
-
-    # Convert to a dict to check the values
-    actual_chemistries = asdict(actual_chemistries)
-
-    # Check that all keys match and have correct values for both dictionaries
-    assert set(expected_chemistries.keys()) == set(actual_chemistries.keys())
-
-    for key in actual_chemistries.keys():
-        assert np.allclose(actual_chemistries[key], expected_chemistries[key])
-
-
-def test_calculate_litter_input_lignin_concentrations(litter_inputs):
-    """Check calculation of lignin concentrations of each plant flow to litter."""
-    from virtual_ecosystem.models.litter.chemistry import (
-        calculate_litter_input_lignin_concentrations,
-    )
-
-    expected_woody = [0.233, 0.545, 0.612, 0.378]
-    expected_concs_above_struct = [0.25011178, 0.25345463, 0.54339369, 0.61992378]
-    expected_concs_below_struct = [0.48590258, 0.56412613, 0.54265483, 0.67810978]
-
-    actual_concs = calculate_litter_input_lignin_concentrations(
-        litter_inputs=litter_inputs,
-    )
-
-    assert np.allclose(actual_concs["woody_lignin"], expected_woody)
-    assert np.allclose(
-        actual_concs["above_structural_lignin"], expected_concs_above_struct
-    )
-    assert np.allclose(
-        actual_concs["below_structural_lignin"], expected_concs_below_struct
-    )
-
-
-def test_calculate_litter_input_nitrogen_ratios(litter_inputs):
-    """Check function to calculate the C:N ratios of input to each litter pool works."""
-    from virtual_ecosystem.models.litter.chemistry import (
-        calculate_litter_input_nitrogen_ratios,
-    )
-
-    expected_c_n_ratios = {
-        "woody_nitrogen": [60.7, 57.9, 73.1, 55.1],
-        "below_metabolic_nitrogen": [
-            20.32269136,
-            22.96676383,
-            26.06473456,
-            19.59251036,
-        ],
-        "below_structural_nitrogen": [
-            101.61345679,
-            114.83381916,
-            130.32367278,
-            97.96255178,
-        ],
-        "above_metabolic_nitrogen": [12.540983, 21.600478, 20.237902, 15.403147],
-        "above_structural_nitrogen": [62.91002, 110.3194, 109.3635, 75.59183],
-    }
-
-    actual_c_n_ratios = calculate_litter_input_nitrogen_ratios(
-        litter_inputs=litter_inputs,
-        struct_to_meta_nitrogen_ratio=LitterConsts.structural_to_metabolic_n_ratio,
-    )
-
-    assert set(expected_c_n_ratios.keys()) == set(actual_c_n_ratios.keys())
-
-    for key in actual_c_n_ratios.keys():
-        assert np.allclose(actual_c_n_ratios[key], expected_c_n_ratios[key])
-
-
-def test_calculate_litter_input_phosphorus_ratios(litter_inputs):
-    """Check function to calculate the C:P ratios of input to each litter pool works."""
-    from virtual_ecosystem.models.litter.chemistry import (
-        calculate_litter_input_phosphorus_ratios,
-    )
-
-    expected_c_p_ratios = {
-        "woody_phosphorus": [856.5, 675.4, 933.2, 888.8],
-        "below_metabolic_phosphorus": [
-            440.4591226,
-            226.94788998,
-            263.23576031,
-            196.40039357,
-        ],
-        "below_structural_phosphorus": [
-            2202.29561299,
-            1134.7394499,
-            1316.17880156,
-            982.00196785,
-        ],
-        "above_metabolic_phosphorus": [286.886303, 107.015923, 241.802298, 136.049497],
-        "above_structural_phosphorus": [
-            1488.595406,
-            580.6433876,
-            1408.378272,
-            610.0666667,
-        ],
-    }
-
-    actual_c_p_ratios = calculate_litter_input_phosphorus_ratios(
-        litter_inputs=litter_inputs,
-        struct_to_meta_phosphorus_ratio=LitterConsts.structural_to_metabolic_p_ratio,
-    )
-
-    assert set(expected_c_p_ratios.keys()) == set(actual_c_p_ratios.keys())
-
-    for key in actual_c_p_ratios.keys():
-        assert np.allclose(actual_c_p_ratios[key], expected_c_p_ratios[key])
-
-
-def test_calculate_nutrient_split_between_litter_pools(
-    dummy_litter_data, litter_inputs
-):
-    """Check the function to calculate the nutrient split between litter pools."""
-    from virtual_ecosystem.models.litter.chemistry import (
-        calculate_nutrient_split_between_litter_pools,
-    )
-
-    expected_meta_c_n = np.array([20.32269136, 22.96676383, 26.06473456, 19.59251036])
-    expected_struct_c_n = np.array([101.6134568, 114.83381915, 130.3236728, 97.9625518])
-
-    actual_meta_c_n, actual_struct_c_n = calculate_nutrient_split_between_litter_pools(
-        input_c_nut_ratio=dummy_litter_data["root_turnover_c_n_ratio"],
-        metabolic_split=litter_inputs.roots_meta_split,
-        struct_to_meta_nutrient_ratio=LitterConsts.structural_to_metabolic_n_ratio,
-    )
-
-    # Standard checks of the produced values
-    assert np.allclose(actual_meta_c_n, expected_meta_c_n)
-    assert np.allclose(actual_struct_c_n, expected_struct_c_n)
-    # Check that expected ratio is actually preserved by the function
-    assert np.allclose(
-        expected_struct_c_n,
-        expected_meta_c_n * LitterConsts.structural_to_metabolic_n_ratio,
-    )
-    # Check that weighted sum of the two new C:N ratios is compatible with the original
-    # C:N ratio
-    assert np.allclose(
-        dummy_litter_data["root_turnover_c_n_ratio"],
-        1
-        / (
-            (litter_inputs.roots_meta_split / actual_meta_c_n)
-            + ((1 - litter_inputs.roots_meta_split) / actual_struct_c_n)
-        ),
-    )
